ENH: LHS sample using fitted parameters
N was not fitting at all due to the discrete "mu" parameter.  The underlying
issue was when optim() tried varying "mu", the output would often not change
and therefore the optimziation would "complete".  The fault is not completely
fixed but somewhat mitigated by reparameterizing N using all continuous
parameters, by replacing "mu" with the "prob" parameter.  However the
optimization still exists prematurely because of the discrete output of the
discrete negative binomial distribution.  Fixing this is not trivial.


git-svn-id: https://innoculant.micro.med.umich.edu/svn/dev/papers/calipro-vs-abc@8333 88ce85bb-8564-4ae6-884f-ab7503230054